// src/app/user/models/feature.model.ts
export class Feature {
    code: string;
    name: string;
    title: string;
    displayOrder: number;
    parent: any;
    isMenuFeature: boolean;
    isToolbarFeature: boolean;
    isAppFeature: boolean;
    isWebFeature: boolean;
    featureType: string;
    featureRoute: string;
    featureWebIconClass: string;
    featureAppIconClass: string;
}